bit more, refactor types

+++ src/18ce0.c
@@ -1,4 +1,6 @@
-int func_18ce0(long long* arg0, long long arg1) {
+#include "types.h"
+
+s32 func_18ce0(u64* arg0, u64 arg1) {
     *arg0 = arg1;
 
     return 0;
