+++ src/4c28.c
@@ -1,7 +1,8 @@
 // #include "include_asm.h"
 #include "include_asm.h"
 #include "structs.h"
+#include "types.h"
 
-int __attribute__((naked)) func_4c28(unk_struct_1188_3* arg0, long long arg1, int arg2, int* arg3) {
+s32 __attribute__((naked)) func_4c28(unk_struct_1188_3* arg0, u64 arg1, s32 arg2, s32* arg3) {
     INCLUDE_ASM("asm/nonmatching", func_4c28);
 }
